% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF.R
\name{cdfSub}
\alias{cdfSub}
\title{Utility function}
\usage{
cdfSub(v, q, alpha1, alpha2, lambda1, lambda2)
}
\arguments{
\item{v}{integration term}

\item{q}{quantile}

\item{alpha1}{parameter value}

\item{alpha2}{parameter value}

\item{lambda1}{parameter value}

\item{lambda2}{parameter value}
}
\description{
Utility function
}

