% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_MoM.R
\name{g}
\alias{g}
\title{first derivative of log likelihood}
\usage{
g(N, muHat, sumLog, alpha)
}
\arguments{
\item{N}{total sample size}

\item{muHat}{sample mean}

\item{sumLog}{sum of log(x_i)}

\item{alpha}{parameter value}
}
\description{
first derivative of log likelihood
}

