% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_MoM.R
\name{gammaFit}
\alias{gammaFit}
\title{Maximum likelihood and method of moments estimates for iid observations of gamma random variables}
\usage{
gammaFit(z, maxIter = 1000, tol = 1e-04)
}
\arguments{
\item{z}{data}

\item{maxIter}{maximum number of iterations (maxIter = 1000 by default)}

\item{tol}{tolerance for convergence (|logLik_k - logLik_{k-1}| < tol)}
}
\description{
This function computes maximum likelihood and method of moments estimates
for iid observations of gamma random variables.
Maximum likelihood estimatation is via a Newton-Raphson routine.
}
\examples{
nx <- 60
ny <- 60
alpha <- 1
lambda <- 4
x <- rgamma(n = nx, shape = alpha, rate = lambda)
y <- rgamma(n = ny, shape = alpha, rate = lambda)
fit <- gammaFit(z = c(x,y))
fit$MLE
fit$MoM
}

