% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE_MoM.R
\name{h}
\alias{h}
\title{second derivative of log likelihood}
\usage{
h(N, muHat, alpha)
}
\arguments{
\item{N}{total sample size}

\item{muHat}{sample mean}

\item{alpha}{parameter value}
}
\description{
second derivative of log likelihood
}

