% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle_point.R
\name{K}
\alias{K}
\title{cumulant generating function (CGF)}
\usage{
K(t, nx, ny, alpha, lambda)
}
\arguments{
\item{t}{CGF argument}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
cumulant generating function (CGF)
}

