% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle_point.R
\name{Kprime}
\alias{Kprime}
\title{first derivative of cumulant generating function (CGF)}
\usage{
Kprime(t, nx, ny, alpha, lambda)
}
\arguments{
\item{t}{CGF argument}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
first derivative of cumulant generating function (CGF)
}

