% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parest.R
\name{parest}
\alias{parest}
\title{Estimate the parameters of the gamma distribution.}
\usage{
parest(x)
}
\arguments{
\item{x}{gamma distributed observations.}
}
\value{
parest gives the estimates of the shape parameter, the rate parameter and and the scale parameter.
}
\description{
Estimate the parameters of the gamma distribution.
}
\examples{
x <- rGamma(1000, shape = 2, rate = 1)
parest(x)

x <- c(1.6, 2.7, 1.4, 2.4, 1.3, 1.8, 2.3, 1.1)
parest(x)
}
\references{
Ye, Z.-S. and Chen, N. (2017), “Closed-form estimators for the gamma distribution derived from likelihood equations,” The American Statistician, 71(2), 177–181.

Louzada, F., Ramos, P. L., and Ramos, E. (2019), “A note on bias of closed-form estimators for the gamma distribution derived from likelihood equations,” The American Statistician, 73(2), 195–199.
}
