% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle_point.R
\name{pgammaDifSaddle}
\alias{pgammaDifSaddle}
\title{Saddlepoint distribution for difference in means, \eqn{\bar{x} - \bar{y}}}
\usage{
pgammaDifSaddle(q, nx, ny, alpha, lambda, lower.tail = TRUE)
}
\arguments{
\item{q}{scalar or vector of quantiles}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
This function calculates the saddlepoint cumulative distribution for 
the difference in means, \eqn{\bar{x} - \bar{y}} of gamma
random variables under the null of equal distributions. Only valid
for \eqn{\bar{x} - \bar{y} \ne 0}.
}
\examples{
nx <- 60
ny <- 60
alpha <- 1
lambda <- 4
q <- seq(-0.2, 0.2, 0.001)
plot(q, pgammaDifSaddle(q = q, nx, ny, alpha, lambda),
     type = "l", ylab = "F(q)")
}

