% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGamma.R
\name{rGamma}
\alias{rGamma}
\title{Generate gamma distributed random variables.}
\usage{
rGamma(n, shape, rate = 1, scale = 1/rate)
}
\arguments{
\item{n}{number of observations.}

\item{shape}{value of the shape parameter (> 0).}

\item{rate}{value of the rate parameter (> 0).}

\item{scale}{value of the scale parameter (> 0).}
}
\value{
rGamma(n, shape, rate, scale) generates n gamma distributed random numbers with parameters
shape and rate.
}
\description{
Generate gamma distributed random variables.
}
\details{
If rate and scale are ommitted, it assumes the default value of 1.
}
\examples{
rGamma(100, shape = 0.5, rate = 3)

rGamma(1000, shape = 2, scale = 2)
}
\references{
Liu, C., Martin, R., and Syring, N. (2017), “Efficient simulation from a gamma distribution with
small shape parameter,” Computational Statistics, 32(4), 1767–1775.
}
