% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle_point.R
\name{saddleEqn}
\alias{saddleEqn}
\title{saddlepoint equation: solve to obtain the saddlepoint \eqn{\hat{t}}}
\usage{
saddleEqn(t, q, nx, ny, alpha, lambda)
}
\arguments{
\item{t}{CGF argument}

\item{q}{quantile}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
saddlepoint equation: solve to obtain the saddlepoint \eqn{\hat{t}}
}

