% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddle_point.R
\name{dgammaDifSaddle}
\alias{dgammaDifSaddle}
\title{Saddlepoint density for difference in means, \eqn{\bar{x} - \bar{y}}}
\usage{
dgammaDifSaddle(z, nx, ny, alpha, lambda)
}
\arguments{
\item{z}{scalar or vector of quantiles}

\item{nx}{sample size of first group}

\item{ny}{sample size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
This function calculates the saddlepoint density for difference in means,
\eqn{\bar{x} - \bar{y}} of observations of gamma random variables
under the null of equality in distributions.
}

