% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF.R
\name{pgammaDif}
\alias{pgammaDif}
\title{True CDF for difference in means \eqn{\bar{x} - \bar{y}}}
\usage{
pgammaDif(q, nx, ny, alpha, lambda)
}
\arguments{
\item{q}{scalar or vector of quantiles}

\item{nx}{sample size of first group}

\item{ny}{samle size of second group}

\item{alpha}{common parameter value}

\item{lambda}{common parameter value}
}
\description{
This function computes the true CDF for the difference in means \eqn{\bar{x} - \bar{y}}
of gamma random variables under the null of equality in distributions.
CDF derived by Klar (2015).
}
\examples{
nx <- 60
ny <- 60
alpha <- 1
lambda <- 4
q <- seq(-0.2, 0.2, 0.001)
plot(q, pgammaDif(q = q, nx, ny, alpha, lambda),
     type = "l", ylab = "F(q)")
}

